Refactor scid-mgr into reusable Rust library crate and thin CLI binary

+++ src/db.rs
@@ -740,7 +740,7 @@ impl ScidDatabaseWrapper {
                         let matches: Vec<crate::position_search::PositionMatch> = gids
                             .into_iter()
                             .map(|gid| crate::position_search::PositionMatch {
-                                game_id: gid as usize,
+                                game_id: gid,
                                 ply: 0,
                             })
                             .collect();
